certificate_system/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
qrcode[pil]>=7.4,<8.0
reportlab>=4.0,<5.0
python-dotenv>=1.0,<2.0
python-json-logger>=2.0,<3.0
Pillow>=10.0,<12.0
gunicorn>=21.2,<23.0
